requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
imblearn
xgboost
catboost
openpyxl
ipykernel
flask
dill
nbformat
dvc
mlflow
python-dotenv
mysql-connector-python
pymysql
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
python-box==6.0.2


-e .